Business-Conversational-AI-main/backend/requirements.txt
fastapi
uvicorn
pandas
google-generativeai
pydantic
sqlalchemy
python-dotenv
